+++ gmx_lib/solvent_equilibration.py
@@ -49,7 +49,7 @@ def run(params, stage):
 	# check convergence
 	print "\t ", params.gmx_engine.convergence_line(program)
 
-	os.remove(traj_out)
-	os.remove(edrfile_out)
+	if os.path.exists(traj_out): os.remove(traj_out)
+	if os.path.exists(edrfile_out): os.remove(edrfile_out)
